requirements.txt
paho-mqtt
zeroconf
aiohttp
aiomqtt